refactor: directory for each type of HTTP message
