+++ src/class/http/HttpTransaction.cpp
@@ -1,7 +1,7 @@
 #include "./HttpTransaction.hpp"
 #include "http/errors/HttpException.hpp"
-#include "http/messages/HttpRequest.hpp"
-#include "http/messages/HttpResponse.hpp"
+#include "http/messages/request/HttpRequest.hpp"
+#include "http/messages/response/HttpResponse.hpp"
 #include <iostream>
 
 HttpTransaction::HttpTransaction() : _request(*this), _response(*this), _last(false) {
