+++ src/class/http/messages/request/HttpRequest.cpp
@@ -1,4 +1,4 @@
-#include "http/messages/HttpRequest.hpp"
+#include "./HttpRequest.hpp"
 #include "http/HttpTransaction.hpp"
 #include "http/messages/HttpMessage.hpp"
 #include "http/types.hpp"
